tests slight cleanup
